Refactor `ProcedureDispatcher` and test files for handler lookup transition.
- Rename `getHandlerForProcedure` to `getHandler` in `ProcedureDispatcher` and update all usages.
- Change `ProcedureDispatcher`'s handler mapping to use procedure classes instead of types for improved flexibility.
- Update test cases in `ProcedureDispatcherTest` to align with the new method and mapping logic.
- Adjust imports in `CLIMain` and `ProcedureDispatcher` for the package name update.
- Enhance inspection profile settings to add a new inspection rule.
